+++ eq-author/src/App/page/Design/QuestionPageEditor/AdditionalInfo.js
@@ -1,12 +1,17 @@
 import React from "react";
+import styled from "styled-components";
 import PropTypes from "prop-types";
 import { propType } from "graphql-anywhere";
-import { flowRight } from "lodash";
+import { get, flowRight } from "lodash";
 import { TransitionGroup } from "react-transition-group";
 
 import WrappingInput from "components/Forms/WrappingInput";
 import RichTextEditor from "components/RichTextEditor";
 import { Field, Label } from "components/Forms";
+
+import { colors } from "constants/theme";
+
+import focusOnNode from "utils/focusOnNode";
 
 import withChangeUpdate from "enhancers/withChangeUpdate";
 
@@ -25,6 +30,35 @@ const contentControls = {
   link: true,
 };
 
+const descriptionControls = {
+  emphasis: true,
+  piping: true,
+  link: true,
+};
+
+const definitionControls = {
+  list: true,
+  emphasis: true,
+  bold: true,
+  link: true,
+};
+
+const guidanceControls = {
+  bold: true,
+  list: true,
+  link: true,
+};
+
+const Paragraph = styled.p`
+  margin: 0 0 1em;
+  background: ${colors.lighterGrey};
+  padding: 0.5em;
+  border-left: 5px solid ${colors.lightGrey};
+`;
+
+const errorMsg = (field, page) =>
+  getErrorByField(field, page.validationErrorInfo.errors);
+
 export const StatelessAdditionalInfo = ({
   page,
   onChange,
@@ -34,9 +68,10 @@ export const StatelessAdditionalInfo = ({
   page: {
     validationErrorInfo: { errors },
   },
+  option,
 }) => (
   <TransitionGroup>
-    {page.additionalInfoEnabled && (
+    {page.additionalInfoEnabled && option === "additionalInfoEnabled" ? (
       <AnswerTransition
         key="additional-info"
         onEntered={() => focusOnElement("additional-info-label")}
@@ -79,7 +114,92 @@ export const StatelessAdditionalInfo = ({
           />
         </MultipleFieldEditor>
       </AnswerTransition>
-    )}
+    ) : null}
+
+    {page.descriptionEnabled && option === "descriptionEnabled" ? (
+      <AnswerTransition
+        key="question-description"
+        onEntered={() => focusOnNode(page.description)}
+      >
+        <RichTextEditor
+          ref={undefined}
+          id="question-description"
+          name="description"
+          label="Question description"
+          multiline
+          value={page.description}
+          onUpdate={onChangeUpdate}
+          controls={descriptionControls}
+          fetchAnswers={fetchAnswers}
+          metadata={get(page, "section.questionnaire.metadata", [])}
+          testSelector="txt-question-description"
+          errorValidationMsg={errorMsg("description", page)}
+        />
+      </AnswerTransition>
+    ) : null}
+
+    {page.definitionEnabled && option === "definitionEnabled" ? (
+      <AnswerTransition
+        key="definition"
+        onEntered={() => focusOnElement("definition-label")}
+      >
+        <MultipleFieldEditor id="definition" label="Question definition">
+          <Paragraph>
+            Only to be used to define word(s) or acronym(s) within the question.
+          </Paragraph>
+          <Field>
+            <Label htmlFor="definition-label">Label</Label>
+            <WrappingInput
+              id="definition-label"
+              name="definitionLabel"
+              data-test="txt-question-definition-label"
+              onChange={onChange}
+              onBlur={onUpdate}
+              value={page.definitionLabel}
+              bold
+              errorValidationMsg={errorMsg("definitionLabel", page)}
+            />
+          </Field>
+          <RichTextEditor
+            id="definition-content"
+            name="definitionContent"
+            label="Content"
+            multiline
+            value={page.definitionContent}
+            onUpdate={onChangeUpdate}
+            controls={definitionControls}
+            fetchAnswers={fetchAnswers}
+            metadata={page.section.questionnaire.metadata}
+            testSelector="txt-question-definition-content"
+            errorValidationMsg={errorMsg("definitionContent", page)}
+          />
+        </MultipleFieldEditor>
+      </AnswerTransition>
+    ) : null}
+
+    {page.guidanceEnabled && option === "guidanceEnabled" ? (
+      <TransitionGroup>
+        <AnswerTransition
+          key="question-guidance"
+          onEntered={() => focusOnNode(page.guidance)}
+        >
+          <RichTextEditor
+            ref={page.guidance}
+            id="question-guidance"
+            name="guidance"
+            label="Include/exclude"
+            multiline
+            value={page.guidance}
+            onUpdate={onChangeUpdate}
+            controls={guidanceControls}
+            fetchAnswers={fetchAnswers}
+            metadata={get(page, "section.questionnaire.metadata", [])}
+            testSelector="txt-question-guidance"
+            errorValidationMsg={errorMsg("guidance", page)}
+          />
+        </AnswerTransition>
+      </TransitionGroup>
+    ) : null}
   </TransitionGroup>
 );
 
@@ -89,6 +209,7 @@ StatelessAdditionalInfo.propTypes = {
   fetchAnswers: PropTypes.func.isRequired,
   page: propType(pageFragment).isRequired,
   onChangeUpdate: PropTypes.func.isRequired,
+  option: PropTypes.string.isRequired,
 };
 
 StatelessAdditionalInfo.fragments = {
